add Distributed locks in order to protect listner

+++ huhu-search-service/src/main/java/top/yeonon/huhusearchservice/listener/BinlogListener.java
@@ -4,6 +4,8 @@ import com.github.shyiko.mysql.binlog.BinaryLogClient;
 import com.github.shyiko.mysql.binlog.event.*;
 import com.google.common.collect.Maps;
 import lombok.extern.slf4j.Slf4j;
+import org.redisson.api.RLock;
+import org.redisson.api.RedissonClient;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -11,6 +13,7 @@ import org.springframework.context.annotation.DependsOn;
 import org.springframework.stereotype.Component;
 import top.yeonon.huhucommon.exception.HuhuException;
 import top.yeonon.huhusearchservice.constant.MysqlConst;
+import top.yeonon.huhusearchservice.constant.RedisConst;
 import top.yeonon.huhusearchservice.listener.handler.EventDataHandler;
 import top.yeonon.huhusearchservice.util.SearchUtils;
 
@@ -19,6 +22,7 @@ import java.io.IOException;
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
 
 /**
  * @Author yeonon
@@ -40,6 +44,8 @@ public class BinlogListener {
 
     private final EventDataHandler userEventDataHandler;
 
+    private final RedissonClient redissonClient;
+
 
     @Value("${spring.datasource.username}")
     private String username;
@@ -54,10 +60,11 @@ public class BinlogListener {
     @Autowired
     public BinlogListener(@Qualifier("questionEventDataHandler") EventDataHandler questionEventDataHandler,
                           @Qualifier("answerEventDataHandler") EventDataHandler answerEventDataHandler,
-                          @Qualifier("userEventDataHandler") EventDataHandler userEventDataHandler) {
+                          @Qualifier("userEventDataHandler") EventDataHandler userEventDataHandler, RedissonClient redissonClient) {
         this.questionEventDataHandler = questionEventDataHandler;
         this.answerEventDataHandler = answerEventDataHandler;
         this.userEventDataHandler = userEventDataHandler;
+        this.redissonClient = redissonClient;
     }
 
     private static Map<Long, TableMapEventData> tablesById = Maps.newConcurrentMap();
@@ -94,24 +101,40 @@ public class BinlogListener {
 
         @Override
         public void onEvent(Event event) {
-            EventType eventType = event.getHeader().getEventType();
+            //获取分布式锁
+            RLock lock = redissonClient.getLock(RedisConst.BINLOG_CHANGE_LOCK);
+            boolean getLock = false;
+            try {
+                if (getLock = lock.tryLock(RedisConst.BINLOG_GET_LOCK_TIMEOUT, TimeUnit.SECONDS)) {
+                    EventType eventType = event.getHeader().getEventType();
+                    switch (eventType) {
+                        case TABLE_MAP:
+                            TableMapEventData tableMapEventData = event.getData();
+                            tablesById.put(tableMapEventData.getTableId(), tableMapEventData);
+                            break;
+                        case EXT_WRITE_ROWS:
+                            handleWriteRowEvent(event);
+                            break;
+                        case EXT_UPDATE_ROWS:
+                            handleUpdateRowEvent(event);
+                            break;
+                        case EXT_DELETE_ROWS:
+                            handleDeleteRowEvent(event);
+                        default:
+                            //ignore
+                    }
+                }
 
-            switch (eventType) {
-                case TABLE_MAP:
-                    TableMapEventData tableMapEventData = event.getData();
-                    tablesById.put(tableMapEventData.getTableId(), tableMapEventData);
-                    break;
-                case EXT_WRITE_ROWS:
-                    handleWriteRowEvent(event);
-                    break;
-                case EXT_UPDATE_ROWS:
-                    handleUpdateRowEvent(event);
-                    break;
-                case EXT_DELETE_ROWS:
-                    handleDeleteRowEvent(event);
-                default:
-                    //ignore
+            } catch (InterruptedException e) {
+                log.error(e.getMessage());
+            } finally {
+                //如果成功获取锁了，最后释放锁
+                //没有成获取则不需要也不能重复释放锁
+                if (getLock) {
+                    lock.unlock();
+                }
             }
+
         }
     }
